Implement basic host store (#69)
Follow up to #60.

Implement a basic `Store` that's able to store the `Blueprint` and load
it after. This let's the plan command move a bit further ahead. Next &
final stop, actually calculating the plan (last TODO in the code).

This required fixing various random bugs, and implement marshaling for
some types. These marshalings are a bit tricky, as the structs have the
interface references, but we must marshal the contrete type, and
unmarshal it back to the interface. I've added tests to cover that.