Plan: more ground work (#74)
This PR moves us closer to having a "plan" command. It is a split of
#73, that was getting too big. It conatis various bits of groundwork,
required for plan to work later on.

- Store
- Add `--store*` CLI options, to supprot storing state either at target
host or localhost.
- Move `Step` to its owns file.
- `File` was fixed to allow 0 on permissions. We need to allow this to
be able to detect a Resource with Remove set together with other
attributes as an error.
- Log handler: for attribute values, if they use ANSI escape code, we
let them out verbatim, instead of escaping it. This is required to allow
showing colored diffs between resource versions.
- Resources
- Add more validations to Resources, that now must have specific yaml
tags, a Remove field an a string field tagged with resonance:"id".
  - Added various new functions, that'll be used later.
  - Added tests.

Follow up to https://github.com/fornellas/resonance/pull/60,
https://github.com/fornellas/resonance/pull/69.