Fix update (#79)
Found the issue: `go get -u`, which we were running until now, was NOT
updating anything on `tools.go`, which has... `staticcheck` (and other
tools). the n-1 version of staticcheck fails wit the hot off the press
go 1.23.0, but staticcheck version n does work. So we update both,
and... it works.

To update things at `tools.go`, we need to do `got get -u -t`.The
documentation for `-t` is deceiving at best:

```
The -t flag instructs get to consider modules needed to build tests of
packages specified on the command line.
```

`tools.go` is definitely not a test... but the imports there are updated
when `-t` is passed...

TLDR, add `-t` to fix this, and also do the update.

Closes #76.