Fix log indentation
The logging handler for console supports indentation, which can be interpreted as a "section".

As of now, indentation can be arbitrarily created, but there's nothing enforcing that ther's any sort of title. This can result in odd outputs.

Eg: instead of :

```
top message
  mid message
    lower message
```

if nothing is loggen on mid, we end up with:

```
top message
    lower message
```

which looks odd. This can happen when for example, mid is only logging on debug, and we're running on info.

This PR (for the most part) addresses that, by changing the interface to "ask for a section title", and enforces the logging of the title as info.

commit-id:6b80a22d