Add .spr.yaml
GitHub is pretty terrible to deal with stacked PRs. I've found https://github.com/ejoffe/spr which seems promissing to help aliviate the pain.

This PR adds the configuration for it, so we can start playing and see if it lives up to the expectation.

I've also added a branch protection rule against any merges to `spr/master/*`, no bypassing, as spr / GH will go banannas if we allow that (and accidentaly merge via the web UI...).

commit-id:fc3c74d2