Simplify id & remove resource fields
All resource types must have both an "id" field, that uniquely identifies the resource at a host, and a "remove" field, to inform whether it should be... removed.

As of now, both fields are mandatory, but they can live anywhere in the struct, and the Id field, requires a special tag.

This PR simplifies that, enforcing instead that all resources have the first 2 fields as such:

```go
type Foo struct {
	FooName string `yaml:"foo_name"`
	Remove bool `yaml:"remove,omitempty"`
}
```

which simplifies things quite a bit.

In particular, when marshaling yamls, the id field will be guaranteed to be the first one, which eases understanding.

commit-id:920d61a3