Resources: Add Satisfies
When comparing some resources, it is not enough to just look "is it the name values". Eg: for a File, if all attributes are the same for 2 resources, then it is the same state. Eg: for APTPackage, if (a) does not require any specific version, but (b) does, then (b) satisfies (a), but (a) does not satisfy (b).

This PR implements this interface, which is required for planning to work.

commit-id:b6dad582