Change 'remove' to 'absent'
Resources must have a field to indicate whether it is there or not. Until now, the word 'remove' has been used. When implementing the plan mechanics, I figured that this term can create confusion in some circunstances.

This PR changes the name 'remove' to 'absent', which when true, indicates... the resource is absent.

commit-id:bff52a7f