Fix console logger indentation for level > info
PR #90 introduced sections and logs as info when a new section is created. However, if the log level is higher (eg: warn), then the section won't be logged, and any higher level log (eg: warn, error) will be indented without a section.

This PR fixes that, by only enabling indentation when the level info is enabled.

commit-id:574d04a7