Add apply command
Follow up to #107 (which added the `plan` command). Now we're _finally_ at the point where things are starting to be functional & testable.

Once this PR merges, we should have all the basic inner workings to calculate all changes, plan and apply them, including all state management (eg: to allow restoring to the original state when we no longer manage a resource).

Getting closer to the first worknig alpha release!

Closes #53.

commit-id:34d8d120