~10x faster & ~700x less memory for the HTTP agent
The HTTP agent is using a yaml library that has proven to be terrible for CPU & memory usage. When testing #159, I found peaks of the agent consuming *~30G* (yes, with a "G") of RAM, a taking ~1min to run. Changing from yaml to go's own json library, we go to well behaved memory usage proportinoal to the ~20M of data we have and ~6s to run.

Here's a memory profile I took, while the agent was in the process of getting to ~30G peak:

![image](https://github.com/user-attachments/assets/f87c0a98-60c1-421f-b19f-4185697e62a0)

@ricardo6142dh this definitely closes the gap to gRPC, but I do expect gRPC to beat JSON on CPU usage in the end.

commit-id:383ceb62