Change Host.ReadFile to stream files
ATM we buffer the whole file contents in memory, which is far from optimal.

This PR changes the interface to return a `io.ReadCloser`, enabling file chunks to be read, without having to put 100% of it in memory. This should both speed up things and prevent OOMs in certain scenarios.

commit-id:b2f4b3f2