Stream AgentClientWrapper.Run stdout/stderr
ATM the agent server runs the command, fully buffers the stdout & stderr in memory, send all of it through the wire, then the agent client copies all of it to the client. This can yield high memory usage for big outputs and is rather slow.

Let's change the implementation, to have the stdout / stderr to be streamed to `Cmd.Stdout` and `Cmd.Stderr` instead, like it already happens for all other host implementations.

commit-id:5d1645a9