Add tests for Docker
We were missing tests for Docker, let's add them.

This caught a bug, where the client's environment was being wrongly copied to the target. To fix this, we're now defining a `DefaultEnv`, and having the tests assert for it. This ensures predictability when running commands, as opposed to the client's environment leaking to the target.

commit-id:b23ce8c3