Deprecate command based Host implementations (#214)
We have a fallback `Host` implementation that relies on shelling out
commands, despite it being hyper slow. While working on #211, it became
clear that it'd be borderline impossible to provide a full shellout
implementation for `Host.Mknod`, due to the absence of a common tool
across various Linux for that.

The solution we're goning here (thanks @rafaelbenvenuti for the input),
is to simply let go of the shellout implementation altogether. The
rationale is:

- It is **extremelly** slow, borderline unusable.
- We now have a fully functioning gRPC agent which is super fast.
- The only scenario where the agent is not supported, is for Linux
GOARCH where there's no support, and this scenario is virtually nil.

This PR removes the command based `Host` implementations, so from now
onwards, it is always agent.

---

**Stack**:
- #159
- #209
- #208
- #215
- #216
- #202
- #211
- #214 ⬅


⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*